chore: move Controller facade to apiserver/facades/controller
