+++ apiserver/facades/controller/controller/mocks/state_mock.go
@@ -10,9 +10,9 @@
 package mocks
 
 import (
+	"github.com/juju/juju/apiserver/facades/controller/controller"
 	reflect "reflect"
 
-	controller "github.com/juju/juju/apiserver/facades/client/controller"
 	charm "github.com/juju/juju/internal/charm"
 	state "github.com/juju/juju/state"
 	gomock "go.uber.org/mock/gomock"
